Create README.md
# ProcessingFlowers
A p5.js program translated  in Processing  and Android.
See https://github.com/Lebo124/Flowersystem for the p5.js files.

Flowersystem

Assignement in Kadenze course "Nature in Code" by Shiffman. Based on the example "03_Faces_interactiveselection". 
In the example "03_Faces_interactiveselection" you can choose between faces and the program makes a new generation based 
on your choice and some mutationvalue. I changed this in a choice between flowers hoh are in generation 0, randomly made 
by the computer on base of 15 genes. This genes take care of the color and the shape of the flowers. 
When the users has made his choices (every flower cab get 1 till 50 points) the user pushes a button and the program will make 
a tree with the flowers of choice. How higher the number (called Fitness), 
how more flowers are on the tree. 
The user pushes a button again and the program makes a new generation by randomly choose a
"mam"and a "dad", who make a child by donating 50% of their dna. 
On every 100 genes there are 15 mutations. This seems a lot. But because we have only 6 flowers 
there is not a very large genespool and like in the real world this means there will be less and less variation. 
The user can try to make more flowers he ore she finds beautiful. I have some ideas to develop the program.


the user can save flowers who he/she likes
the user can choose which genes to keep in the next generation

New ideas are welcome. Also contributions to the code.